Testing support
Using practicalmeteor:mocha and dispatch:mocha-phantomjs (for CI only).

dispatch:mocha-phantom-js has issues running on Windows.  One of them is
that it does not work at all
https://github.com/DispatchMe/meteor-phantomjs-tests/issues/1 and another
is the server side tests are not executed when practicalmeteor:mocha is
added https://github.com/DispatchMe/meteor-mocha-phantomjs/issues/10

Set the test port to 3100 for npm test to allow concurrent runs of
meteor application and the tests.