Testing support with mocha, chai and sinon
Using practicalmeteor:mocha and dispatch:mocha-phantomjs (for CI only).

dispatch:mocha-phantom-js has issues running on Windows.  One of them is
that it does not work at all
https://github.com/DispatchMe/meteor-phantomjs-tests/issues/1 and another
is the server side tests are not executed when practicalmeteor:mocha is
added https://github.com/DispatchMe/meteor-mocha-phantomjs/issues/10

Set the test port to 3100 for meteor npm test to allow concurrent runs of
meteor application and the tests.

Remove practicalmeteor:mocha during CI tests

As noted in
https://github.com/xolvio/meteor-jasmine/issues/334#issuecomment-208725856

It was causing an error to appear in the logs.