Build script changes
Gulpfile is relocated to .gulp/gulpfile.js to totally hide it from Meteor.
gulpfile.js on the root just wraps it

ESDoc is replaced with JSDoc3 to allow documentation of non-ES6
constructs.  Configured jsdoc to render source and line numbers.

Used gulp-prettify to do the html prettify

Enabled tidy of .htmllintrc.