Sample collection
This provides a sample MongoDB collection with methods protected using
mdg:validated-method.  It also provides an example of implementation
visible only on the server side.

It also attaches a schema to ensure the data is consistent on the Mongo
database.

It separates publications into a separate folder.

Indices are created as part of ensureIndex.js to improve performance.

A unit test is also provided.